refactor: extract DashboardMarketCard to web/components/dashboard/
Moves the community market card to a generic, reusable component at
dashboard/dashboard-market-card.tsx for use beyond sports dashboards.
